Added StringUtils for string manipulation methods
