+++ report-api/src/test/java/no/sikt/nva/data/report/api/fetch/FetchDataReportTest.java
@@ -5,8 +5,8 @@ import static java.lang.String.valueOf;
 import static no.sikt.nva.data.report.api.fetch.CustomMediaType.TEXT_CSV;
 import static no.sikt.nva.data.report.api.fetch.CustomMediaType.TEXT_PLAIN;
 import static no.sikt.nva.data.report.api.fetch.formatter.ExpectedCsvFormatter.generateTable;
-import static no.sikt.nva.data.report.api.fetch.formatter.ResultUtils.extractDataLines;
-import static no.sikt.nva.data.report.api.fetch.formatter.ResultUtils.sortResponse;
+import static no.sikt.nva.data.report.api.fetch.formatter.ResultSorter.extractDataLines;
+import static no.sikt.nva.data.report.api.fetch.formatter.ResultSorter.sortResponse;
 import static nva.commons.apigateway.GatewayResponse.fromOutputStream;
 import static org.apache.http.HttpHeaders.ACCEPT;
 import static org.junit.jupiter.api.Assertions.assertEquals;
