+++ report-api/src/test/java/no/sikt/nva/data/report/api/fetch/formatter/ResultSorter.java
@@ -1,31 +1,30 @@
 package no.sikt.nva.data.report.api.fetch.formatter;
 
+import static no.sikt.nva.data.report.api.fetch.formatter.StringUtils.buildString;
+import static no.sikt.nva.data.report.api.fetch.formatter.StringUtils.printAsString;
+import static no.sikt.nva.data.report.api.fetch.formatter.StringUtils.scanData;
 import static no.sikt.nva.data.report.api.fetch.testutils.generator.PublicationHeaders.CONTRIBUTOR_IDENTIFIER;
 import static no.sikt.nva.data.report.api.fetch.testutils.generator.PublicationHeaders.PUBLICATION_ID;
+import static nva.commons.core.StringUtils.EMPTY_STRING;
 import com.google.common.net.MediaType;
 import java.io.IOException;
 import java.io.StringReader;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
-import java.util.Scanner;
 import no.sikt.nva.data.report.api.fetch.CustomMediaType;
+import no.sikt.nva.data.report.api.fetch.formatter.StringUtils.ScanningResult;
 import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVParser;
-import org.apache.commons.csv.CSVPrinter;
 import org.apache.commons.csv.CSVRecord;
 
-public class ResultUtils {
+public class ResultSorter {
 
     private static final int RESULT_HEADER_LAST_INDEX = 1;
     private static final int RESULT_ENDING_FORMATTED_LINE = 1;
     private static final String COLUMN_SPLIT_REGEX = "\\|";
-    private static final String LINE_BREAK = "\n";
-    private static final String EMPTY_STRING = "";
-    private static final String DELIMITER = ",";
 
-    private ResultUtils() {
+    private ResultSorter() {
         // NO-OP
     }
 
@@ -64,26 +63,6 @@ public class ResultUtils {
         return csvRecords;
     }
 
-    private static String printAsString(CSVFormat format, CSVParser csvParser, ArrayList<CSVRecord> recordList)
-        throws IOException {
-        var stringBuilder = new StringBuilder();
-        try (var printer = new CSVPrinter(stringBuilder, format)) {
-            printer.printRecord(csvParser.getHeaderMap().keySet());
-            for (CSVRecord record : recordList) {
-                if (Arrays.stream(record.values()).allMatch(EMPTY_STRING::equals)) {
-                    addNumberOfDelimiters(stringBuilder, record.size() - 1);
-                    continue;
-                }
-                printer.printRecord(record);
-            }
-        }
-        return stringBuilder.toString();
-    }
-
-    public static StringBuilder addNumberOfDelimiters(StringBuilder stringBuilder, int numberOfDelimiters) {
-        return stringBuilder.append(DELIMITER.repeat(Math.max(0, numberOfDelimiters)));
-    }
-
     private static List<String> sortDataLines(ScanningResult scanningResult) {
         var dataLines = scanningResult.lines().subList(0, scanningResult.lines().size() - 1);
         dataLines.sort(Comparator.comparing((String line) -> {
@@ -94,34 +73,5 @@ public class ResultUtils {
             return columns.length > 2 ? columns[2].trim() : EMPTY_STRING;
         }));
         return dataLines;
-    }
-
-    private static String buildString(ScanningResult scanningResult, List<String> dataLines) {
-        var sortedData = new StringBuilder(scanningResult.header())
-                             .append(LINE_BREAK)
-                             .append(scanningResult.separator());
-
-        for (String line : dataLines) {
-            sortedData.append(LINE_BREAK).append(line);
-        }
-        sortedData.append(LINE_BREAK).append(scanningResult.lines().getLast());
-        sortedData.append(LINE_BREAK);
-        return sortedData.toString();
-    }
-
-    private static ScanningResult scanData(String data) {
-        var scanner = new Scanner(data);
-        var header = scanner.nextLine();
-        var separator = scanner.nextLine();
-        var lines = new ArrayList<String>();
-
-        while (scanner.hasNextLine()) {
-            lines.add(scanner.nextLine());
-        }
-        return new ScanningResult(header, separator, lines);
-    }
-
-    private record ScanningResult(String header, String separator, ArrayList<String> lines) {
-
     }
 }
